refactor: modularize ETSformer;
