--- pypots/imputation/etsformer/modules/__init__.py
@@ -1,6 +1,0 @@
-"""
-
-"""
-
-# Created by Wenjie Du <wenjay.du@gmail.com>
-# License: BSD-3-Clause
